The goal of this branch is to provide an experimental support for exhaustive simulation.
It relies on the RTD accessor generated in builder from commit
#77d2afb12fc60b2c7c7869f8297f7db2429beec8 from October 9.

It is actually an extension of the ConcurrentEngine that is called when
the "Do Exhaustive Simulation" check button from the advanced tab is
checked.

It results in a dot file where State label contains a crude
representation of the model state and the transition a representation of
the events (MSE) that occurred from two states.

It will eventually be modified to use a better trace format but can
already be used for some demo. This is a join effort with the recent
branching of timesquare with CADP.

This PR contains:
 * small changes in the CCSLSolver to ease the reset of solver data
 * extension of the Launcher to take into account the "Do Exhaustive
Simulation" button
 * a set of classes to store the state of both the MoCC and the RTD
 * a new execution engine that runs until no new states have to be
visited (if th state space if finite, and run infinitely otherwise)